// case-study-furama/src/app/model/facility.ts
export interface Facility {
  id?: number;
  img?: string;
  name?: string;
  area?: number;
  cost?: number;
  maxPeople?: number;
  rentType?: number;
  facilityType?: number;
  standardRoom?: string;
  descriptionOtherConvenience?: string;
  poolArea?: number;
  numberOfFloors?: number;
  facilityFree?: string;
}